v2/requirements.txt
praw
arxiv
xmltodict
pandas
pytest
scipy
doxygen
python-dotenv

